fix: rename Dialect.ts import to Dialect
